requirements.txt
functions-framework==3.*
requests==2.*
google-cloud-bigquery==3.*
google-cloud-secret-manager==2.*
